# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ParkNow - Parking Reservation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  user: "user",
  supervisor: "supervisor",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface ParkingLot {
  id: number;
  ownerId: number;
  name: string;
  location: string;
  totalSpots: number;
  lat: number;
  lng: number;
  createdAt: string;
}

export interface ParkingLotWithStats {
  id: number;
  ownerId: number;
  name: string;
  location: string;
  totalSpots: number;
  lat: number;
  lng: number;
  createdAt: string;
  availableSpots: number;
  reservedSpots: number;
  occupiedSpots: number;
  /** @nullable */
  distance?: number | null;
}

export type SpotStatus = (typeof SpotStatus)[keyof typeof SpotStatus];

export const SpotStatus = {
  available: "available",
  reserved: "reserved",
  occupied: "occupied",
} as const;

export interface Spot {
  id: number;
  lotId: number;
  spotNumber: string;
  status: SpotStatus;
}

export interface ParkingLotWithSpots {
  id: number;
  ownerId: number;
  name: string;
  location: string;
  totalSpots: number;
  lat: number;
  lng: number;
  createdAt: string;
  spots: Spot[];
}

export interface CreateLotBody {
  name: string;
  location: string;
  totalSpots: number;
  lat: number;
  lng: number;
}

export type UpdateSpotBodyStatus =
  (typeof UpdateSpotBodyStatus)[keyof typeof UpdateSpotBodyStatus];

export const UpdateSpotBodyStatus = {
  available: "available",
  reserved: "reserved",
  occupied: "occupied",
} as const;

export interface UpdateSpotBody {
  status: UpdateSpotBodyStatus;
}

export interface CreateReservationBody {
  spotId: number;
  startTime: string;
  endTime: string;
}

export type ReservationWithDetailsStatus =
  (typeof ReservationWithDetailsStatus)[keyof typeof ReservationWithDetailsStatus];

export const ReservationWithDetailsStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface ReservationWithDetails {
  id: number;
  userId: number;
  spotId: number;
  startTime: string;
  endTime: string;
  status: ReservationWithDetailsStatus;
  spotNumber: string;
  lotName: string;
  lotLocation: string;
  lotLat: number;
  lotLng: number;
  userName: string;
  createdAt: string;
}

export interface DashboardStats {
  totalSpots: number;
  availableSpots: number;
  reservedSpots: number;
  occupiedSpots: number;
  totalReservations: number;
  activeReservations: number;
  totalLots: number;
}

export type GetLotsParams = {
  /**
   * @nullable
   */
  lat?: number | null;
  /**
   * @nullable
   */
  lng?: number | null;
};